Possibly fixed view window issue
The computer border was not wide enough, so I extended it. I could see where the need for extra was, so I extended it to that point, hopefully it works in an itch build. Id not, I'll create the asset that could fix it. (trying to avoid extra work if it's not needed)